+++ app/src/main/java/com/inky/fitnesscalendar/data/measure/Speed.kt
@@ -8,11 +8,17 @@ import kotlinx.parcelize.Parcelize
 
 @Parcelize
 @JvmInline
-value class Speed(val metersPerSecond: Double) : Parcelable, ContextFormat {
+value class Speed(val metersPerSecond: Double) : Parcelable, ContextFormat, Comparable<Speed> {
     val kmh get() = metersPerSecond * 3.6
 
-    override fun formatWithContext(context: Context) = context.getString(R.string.x_kmh, "%.1f".format(kmh))
+    override fun formatWithContext(context: Context) =
+        context.getString(R.string.x_kmh, "%.1f".format(kmh))
+
+    override fun compareTo(other: Speed) = this.metersPerSecond.compareTo(other.metersPerSecond)
 
     operator fun minus(other: Speed) =
         Speed(metersPerSecond = this.metersPerSecond - other.metersPerSecond)
 }
+
+fun Double.kmh() = Speed(metersPerSecond = this / 3.6)
+fun Int.kmh() = this.toDouble().kmh()